+++ pages/formula.php
@@ -73,6 +73,7 @@ if($form[0]['ingredient']){
                         <a href="#" rel="tip" class="text-secondary-emphasis" data-bs-placement="right" title="<?= $cat_details['description'] ?>">
                           <?= ucfirst($meta['catClass']) ?>
                         </a>
+                        <div id="max_usage"></div>
                       </span>
                     </h5>
                   </div>